backend/video-gen-api/requirements.txt
aiohttp==3.10.7
anyio==4.4.0
attrs==24.2.0
ffmpeg==1.4
google-api-core>=2.20.0,<3.0.0
google-auth>=2.35.0
google-cloud-core>=2.4.1
google-cloud-storage>=2.18.2
grpcio-tools>=1.51.1
moviepy==1.0.3
numpy==1.24.3
openai==0.28.0
opencv-python-headless==4.7.0.72
protobuf>=3.19.5,<6.0.0
pydantic==2.8.2
requests==2.32.3
Pillow==9.5.0
Flask==3.0.0
